+++ packages/svelte/src/index.ts
@@ -34,6 +34,7 @@ export { default as Dialog } from "./Dialog.svelte";
 export { default as Toast } from "./Toast.svelte";
 export { default as Table } from "./Table.svelte";
 export { default as Timeline } from "./Timeline.svelte";
+export { default as Steps } from "./Steps.svelte";
 export { default as Progress } from "./Progress.svelte";
 export { default as Slider } from "./Slider.svelte";
 export { default as ColorPicker } from "./ColorPicker.svelte";
@@ -101,6 +102,7 @@ export {
 	XtyleToast,
 	XtyleTable,
 	XtyleTimeline,
+	XtyleSteps,
 	XtyleProgress,
 	XtyleSlider,
 	XtyleColorPicker,