+++ packages/xtyle/src/css/index.ts
@@ -9,6 +9,7 @@ export { iconCss } from "./components/icon.js";
 export { imageCss } from "./components/image.js";
 export { carouselCss } from "./components/carousel.js";
 export { timelineCss } from "./components/timeline.js";
+export { stepsCss } from "./components/steps.js";
 export { parallaxCss } from "./components/parallax.js";
 export { heroCss } from "./components/hero.js";
 export { barCss } from "./components/bar.js";